WIP flowrider refactoring
