+++ flowrider/metaflow_runner.py
@@ -19,8 +19,72 @@ logger = logging.getLogger(__file__)
 SUFFIXES = [".py", ".yaml", ".md", ".txt", ".env.shared"]
 
 
+__all__ = ["run_flow_from_config"]
+
+
+def run_flow_from_config(path: str, tag: str) -> int:
+    """Run flow parameterised by YAML config file.
+
+    Runs flow at `{SRC_DIR}/pipeline/{path}.py` with config
+    from `{SRC_DIR}/config/pipeline/{path}_{tag}.yaml`.
+
+    Args:
+        path:
+        tag:
+
+    Returns:
+        Metaflow run ID
+    """
+
+    # Package source path
+    src_dir = SRC_DIR
+    # Base project path
+    project_dir = src_dir.parent
+
+    # Path to flow
+    flow_path = (src_dir / "pipeline" / path).with_suffix(".py")
+    # Path to flow config
+    config_path = (src_dir / "config" / "pipeline" / f"{path}_{tag}").with_suffix(
+        ".yaml"
+    )
+
+    # Parse config and add run id param
+    config = parse_config(config_path)
+    config["flow_kwargs"]["run-id-file"] = str(
+        config_path.parent / f"{path}_{tag}.run_id"
+    )
+    config["preflow_kwargs"]["package-suffixes"] = ",".join(SUFFIXES)
+    logging.info(config)
+
+    # BUNDLE
+    tmp_path = bundle(
+        project_dir,
+        flow_path,
+        SUFFIXES,
+    )
+    os.chdir(tmp_path)
+    logging.info(f"BUNDLED: [{project_dir},{flow_path}] -> {tmp_path}")
+
+    run_id = execute_flow((tmp_path / path).with_suffix(".py"), **config)
+
+    return run_id
+
+
 def _serialise(x):
     return x if isinstance(x, str) else json.dumps(x)
+
+
+def _parse_options(options):
+    # return t.compose_left(
+    #     t.itemmap(lambda item: (f"--{item[0]}", _serialise(item[1]))),
+    #     lambda x: x.items(),
+    #     chain.from_iterable,
+    #     t.map(quote),
+    # )(options)
+    def foo(item):
+        return quote(lambda item: (f"--{item[0]}", _serialise(item[1])))
+
+    return (foo(option) for option in options)
 
 
 def execute_flow(flow_file: Path, preflow_kwargs: dict, flow_kwargs: dict) -> int:
@@ -80,50 +144,3 @@ def execute_flow(flow_file: Path, preflow_kwargs: dict, flow_kwargs: dict) -> in
         with open(run_id_file, "r") as f:
             run_id = int(f.read())
         return run_id
-
-
-def run_flow_from_config(path: str, tag: str) -> int:
-    """Run flow parameterised by YAML config file.
-
-    Runs flow at `{SRC_DIR}/pipeline/{path}.py` with config
-    from `{SRC_DIR}/config/pipeline/{path}_{tag}.yaml`.
-
-    Args:
-        path:
-        tag:
-
-    Returns:
-    """
-
-    # Package source path
-    src_dir = SRC_DIR
-    # Base project path
-    project_dir = src_dir.parent
-
-    # Path to flow
-    flow_path = (src_dir / "pipeline" / path).with_suffix(".py")
-    # Path to flow config
-    config_path = (src_dir / "config" / "pipeline" / f"{path}_{tag}").with_suffix(
-        ".yaml"
-    )
-
-    # Parse config and add run id param
-    config = parse_config(config_path)
-    config["flow_kwargs"]["run-id-file"] = str(
-        config_path.parent / f"{path}_{tag}.run_id"
-    )
-    config["preflow_kwargs"]["package-suffixes"] = ",".join(SUFFIXES)
-    logging.info(config)
-
-    # BUNDLE
-    tmp_path = bundle(
-        project_dir,
-        flow_path,
-        SUFFIXES,
-    )
-    os.chdir(tmp_path)
-    logging.info(f"BUNDLED: [{project_dir},{flow_path}] -> {tmp_path}")
-
-    run_id = execute_flow((tmp_path / path).with_suffix(".py"), **config)
-
-    return run_id
